restore-sql-server-transaction-logs/Function/requirements.txt
functions-framework==3.5.0
google-api-python-client==2.110.0
google-cloud-storage==2.14.0
google-cloud-error-reporting==1.10.0